requirements.txt
django
gunicorn
django-heroku
setuptools
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
oauth2client
asyncio
async-timeout
https://github.com/Rapptz/discord.py/archive/rewrite.zip